保存报警和变频参数到flash，设备断电数据不丢失，再上电后不需要再手动设置

+++ HARDWARE/LCD/gui.c
@@ -68,8 +68,8 @@ uint8_t limit_temp3_minvalue=0;
 uint8_t limit_temp485_maxvalue=50;//485温度高限值
 uint8_t limit_temp485_minvalue=0;
 
-uint16_t limit_rh_maxvalue=70;//湿度高限值
-uint16_t limit_rh_minvalue=50;
+uint8_t limit_rh_maxvalue=70;//湿度高限值
+uint8_t limit_rh_minvalue=50;
